Add TensorFlow GPU v1.8 during the setup of the virtual env (#9) (#59)
The selection of the regular TensorFlow package or GPU-
accelerated TensorFlow was added into the setup_linux
script.
Although no error is produced if both packages are installed,
the GPU package won't work if the regular package is installed
as well. Therefore, both packages were removed from the conda
environment file.
In the setup_linux script, the selection of the TF package is
done after the environment is created, leaving the regular
release as the default package. Two guards were added for the
creation and update of the conda environment to verify if their
execution was successful or else exit the script with an error
message.
The TF version was set to 1.8 which is the most recent release
that works for both the GPU and regular version.
As a side note, the prettytensor package removed from the conda
environment has the regular TF package as a dependency, and
prettytensor is currently not used in the rllab project, so
it was replaced by the selection of the regular TF package
in the setup script.